Respect GitHub availability when assigning reviews (#365)
* Respect GitHub availability when assigning reviews

GitHub's user status feature allows users to indicate their status and
availability on GitHub. This allows a user to indicate that they have
limited availability and should not be contacted or assigned pending
work. While Cody has its own concept of code review availability, it
would also be good to respect the user's status in GitHub itself, as
this status can also be used by other systems and so would offer a
unified way for a GitHub user to indicate their availability.

* fix rubocop lints

* update github_schema.json